requirements.txt
torch
torchvision
Flask
Pillow
